+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Fixed
+- Removed anonymous parameter from `ProcPtr` type declaration.
+
 ## [2.0.2] - 2017-07-22
 ### Fixed
 - Improve docs.